+++ test/res.jsonp.js
@@ -47,7 +47,7 @@ describe('res', function(){
       request(app)
       .get('/?callback[a]=something')
       .end(function(err, res){
-        res.headers.should.have.property('content-type', 'application/json');
+        res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
         res.text.should.equal('{"count":1}');
         done();
       })
@@ -129,7 +129,7 @@ describe('res', function(){
 
       request(app)
       .get('/')
-      .expect('Content-Type', 'application/vnd.example+json')
+      .expect('Content-Type', 'application/vnd.example+json; charset=utf-8')
       .expect(200, '{"hello":"world"}', done);
     })
 
@@ -158,7 +158,7 @@ describe('res', function(){
         request(app)
         .get('/')
         .end(function(err, res){
-          res.headers.should.have.property('content-type', 'application/json');
+          res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
           res.text.should.equal('null');
           done();
         })
@@ -176,7 +176,7 @@ describe('res', function(){
         request(app)
         .get('/')
         .end(function(err, res){
-          res.headers.should.have.property('content-type', 'application/json');
+          res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
           res.text.should.equal('["foo","bar","baz"]');
           done();
         })
@@ -194,7 +194,7 @@ describe('res', function(){
         request(app)
         .get('/')
         .end(function(err, res){
-          res.headers.should.have.property('content-type', 'application/json');
+          res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
           res.text.should.equal('{"name":"tobi"}');
           done();
         })
@@ -261,7 +261,7 @@ describe('res', function(){
       .get('/')
       .end(function(err, res){
         res.statusCode.should.equal(201);
-        res.headers.should.have.property('content-type', 'application/json');
+        res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
         res.text.should.equal('{"id":1}');
         done();
       })
@@ -280,7 +280,7 @@ describe('res', function(){
       .get('/')
       .end(function(err, res){
         res.statusCode.should.equal(201);
-        res.headers.should.have.property('content-type', 'application/json');
+        res.headers.should.have.property('content-type', 'application/json; charset=utf-8');
         res.text.should.equal('{"id":1}');
         done();
       })